# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The `\"detailed\"` error envelope every handler renders, and what OpenAPI advertises.\n\n`details` is present only on a 422, where it holds one entry per offending field.",
        "properties": {
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field validation failures. Present only on a 422.",
            "title": "Details"
          },
          "error_code": {
            "description": "A stable machine readable code, such as `NOT_FOUND`.",
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "description": "A human readable summary safe to surface to a caller.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "description": "The request id, echoed in the `X-Request-ID` header.",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "The HTTP status code, repeated in the body.",
            "title": "Status",
            "type": "integer"
          },
          "success": {
            "default": false,
            "description": "Always `false` on an error response.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "status",
          "message",
          "request_id",
          "error_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ValidationErrorDetail": {
        "description": "One offending field in a 422, carrying its location and reason but never its input.",
        "properties": {
          "field": {
            "description": "Dotted path to the offending field, or `_root` for the body itself.",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Why the value was rejected.",
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "The pydantic error type, such as `missing` or `string_type`.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ValidationErrorDetail",
        "type": "object"
      },
...
      "WorkspaceOutput": {
        "description": "One non-sensitive root output of a workspace's current state.",
        "properties": {
          "detailed_type": {
            "title": "Detailed Type"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "WorkspaceOutput",
        "type": "object"
      },
      "WorkspaceOutputs": {
        "description": "A workspace's non-sensitive outputs, as remote state sharing hands them to another workspace.\n\nSensitive outputs never appear with a value; only their names are listed, so a\nconsumer can tell a withheld output from a missing one.",
        "properties": {
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceOutput"
            },
            "title": "Outputs",
            "type": "array"
          },
          "sensitive_output_names": {
            "items": {
              "type": "string"
            },
            "title": "Sensitive Output Names",
            "type": "array"
          },
          "state_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Version Id"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id"
        ],
        "title": "WorkspaceOutputs",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Workspaces, variables, config versions and runs, as the gateway routes them.",
    "title": "WebbPulse Terraform control plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/outputs": {
      "get": {
        "description": "The non-sensitive outputs of a workspace's current state.\n\nA run's API token reads only a workspace that is its own, shares globally, or\nnames the run's workspace as a consumer; anything else is a 403\n`REMOTE_STATE_NOT_SHARED`. A workspace with no state yet answers an empty list.",
        "operationId": "read_workspace_outputs_api_v1_workspaces__workspace_id__outputs_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 4,
              "pattern": "^ws-[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOutputs"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          }
        },
        "summary": "Read Workspace Outputs",
        "tags": [
          "workspaces"
        ]
      }
    },
...
  }
}